+++ src/main/java/shop/tryit/domain/order/entity/Order.java
@@ -19,7 +19,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shop.tryit.domain.common.BaseTimeEntity;
 import shop.tryit.domain.member.Member;
-import shop.tryit.domain.order.OrderStatus;
 
 @Entity
 @Getter
