+++ src/main/java/shop/tryit/domain/order/entity/OrderStatus.java
@@ -1,4 +1,4 @@
-package shop.tryit.domain.order;
+package shop.tryit.domain.order.entity;
 
 public enum OrderStatus {
     ORDER, CANCEL
